+++ app/src/main/java/com/hmh/hamyeonham/presentation/achiev/AchievFragment.kt
@@ -1,4 +1,4 @@
-package com.hmh.hamyeonham.presentation.main
+package com.hmh.hamyeonham.presentation.achiev
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.hmh.hamyeonham.R
 
-class MypageFragment : Fragment() {
+class AchievFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_mypage, container, false)
+        return inflater.inflate(R.layout.fragment_achiev, container, false)
     }
 }
