+++ app/src/main/java/com/hmh/hamyeonham/presentation/main/MainActivity.kt
@@ -16,18 +16,11 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun initNavigation() {
-        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
+        val navHostFragment =
+            supportFragmentManager.findFragmentById(R.id.nav_host_fragment) as NavHostFragment
         val navController = navHostFragment.navController
 
         val bottomNavView = findViewById<BottomNavigationView>(R.id.btn_main)
         bottomNavView.setupWithNavController(navController)
-//        val navController =
-//            supportFragmentManager.findFragmentById(
-//                androidx.navigation.fragment.R.id.nav_host_fragment_container,
-//            )?.findNavController()
-//        navController?.let {
-//            val bottomNav = findViewById<BottomNavigationView>(R.id.btn_main)
-//            bottomNav.setupWithNavController(navController)
-//        }
     }
 }
