+++ app/src/main/java/com/hmh/hamyeonham/presentation/main/MainViewModel.kt
@@ -2,5 +2,4 @@ package com.hmh.hamyeonham.presentation.main
 
 import androidx.lifecycle.ViewModel
 
-class MainViewModel : ViewModel() {
-}
+class MainViewModel : ViewModel()
