+++ app/src/main/java/com/hmh/hamyeonham/presentation/mypage/MypageFragment.kt
@@ -1,4 +1,4 @@
-package com.hmh.hamyeonham.presentation.main
+package com.hmh.hamyeonham.presentation.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.hmh.hamyeonham.R
 
-class UsagestatFragment : Fragment() {
+class MypageFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_usagestat, container, false)
+        return inflater.inflate(R.layout.fragment_mypage, container, false)
     }
 }
