+++ app/src/main/res/menu/menu_main.xml
@@ -12,5 +12,4 @@
         android:id="@+id/mypageFragment"
         android:icon="@drawable/ic_launcher_foreground"
         android:title="my page" />
-
 </menu>
