+++ app/src/main/res/navigation/nav_main.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!-- res/navigation/nav_graph.xml -->
+<?xml version="1.0" encoding="utf-8"?><!-- res/navigation/nav_graph.xml -->
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -7,17 +6,20 @@
     app:startDestination="@id/usagestatFragment">
     <fragment
         android:id="@+id/achievFragment"
-        android:name="com.hmh.hamyeonham.presentation.main.AchievFragment"
-        android:label="Achiev Fragment" />
+        android:name="com.hmh.hamyeonham.presentation.achiev.AchievFragment"
+        android:label="Achiev Fragment"
+        tools:layout="@layout/fragment_achiev" />
 
     <fragment
         android:id="@+id/usagestatFragment"
-        android:name="com.hmh.hamyeonham.presentation.main.UsagestatFragment"
-        android:label="Usagestat Fragment" />
+        android:name="com.hmh.hamyeonham.presentation.usagestat.UsagestatFragment"
+        android:label="Usagestat Fragment"
+        tools:layout="@layout/fragment_usagestat" />
 
     <fragment
         android:id="@+id/mypageFragment"
-        android:name="com.hmh.hamyeonham.presentation.main.MypageFragment"
-        android:label="Mypage Fragment" />
+        android:name="com.hmh.hamyeonham.presentation.mypage.MypageFragment"
+        android:label="Mypage Fragment"
+        tools:layout="@layout/fragment_mypage" />
 
 </navigation>
